DRF/requirements.txt
django
djangorestframework
django-cors-headers
wheel
pycuda
numpy
